fix: fix types and refactoring interface
